harden the tests to be more robust
Sometimes the first test failed because of a timeout, due to the
overhead of loading AssetGraph-builder for the first time. It might
be because I'm running these tests on a slow VPS, but CI servers
could very well be as un-beefy, so we might aswell solve it once
and for all.

As the async calls is not canceled on a timeout, it caused a the
next test to fail, as the output from the timedout test would land
in the test/build folder before the second test completed and
after the afterEach hook cleaned it. So I started writing each
test result to a different subfolder in the test/build folder to
avoid cascading of test failures in the future.